Made classes strategy more explicit by using an OrderedDict to preserve binding order.
This was also done in order to give the class dataset a linear memory stucture.
Theoretically, this should make classes generated have the memory addresses for
their children directly attached to the class object in a linear order.
***I haven't done my due diligence yet and tested this.***

The _validate_bindings function was also added to modularize the format for use within
the functions strategy.